Fix buffer overflow in unit->Data.Move.Path[]
When saving a game, save unit->Data.Move only if the current order
seems to use that.  In particular, don't save it when the action
of the current order is UnitActionDie (which doesn't use any data),
nor when the unit hasn't yet started executing the current order.
This fix is incomplete in that it never saves unit->Data.Move for
UnitActionResource, even though that does use it for some values
of unit->SubAction; but this is not a new bug.

When loading a game, raise an error if the path data is too long to
fit in unit->Data.Move.Path[].  This avoids memory corruption that
could have security implications.  However, if the current action is
UnitActionDie, which wouldn't use unit->Data.Move in any case, then
ignore any path data so that games saved before this patch can still
be loaded.

Patch #7070 "Fix buffer overflow in unit->Data.Move.Path[]" by 
Kalle Olavi Niemitalo.