Patch #7111 "Align info labels in map-browser screens"
Line up the left sides of the labels in which Bos Wars shows
information about the selected map: move some of them a little to the
right.  This leaves more space for translations of "File:".

This change affects the map browsers displayed via the Start Game
command or via the Start Editor -> Load Map command.

Patch by Kalle Olavi Niemitalo.