Fix assertion failure in heli death animation
units/heli/ico_heli.png has 10 rows and 8 columns of frames.
The columns correspond to NumDirections = 8.  Because of how
DrawUnitType works, the frame numbers in animations must then
be between 0 and 49, as if there were only 5 directions.
The death animation however used frames 30, 35, 45, and 50; the
out-of-range value caused an assertion failure in debug builds
if a heli was shot down while it happened to be heading northwest.
Fix by using frames 30, 35, 40, and 45 instead.  Frame 40 was
previously unused, and in the PNG, it seems part of the death
sequence.

The failing assertion was added in trunk r9653,
from <http://savannah.nongnu.org/bugs/?28728>.

Reported by llearch. Fixed by Kalle Olavi Niemitalo.