Rearrange Cancel and OK buttons in patch editor menus
Move the Cancel and OK buttons side by side in the patch editor menus,
like they already are in the map editor menus and in some others.
(The OK buttons are not actually labeled "OK".)

It might be more comfortable if the OK button came before the Cancel
button in the tab order, but I think users will be surprised if the
tab order goes from right to left.