Say if the selected image is already used in a patch
In the Create New Patch menu, if the user selects an image that is
already used in some patch type, then display a label saying so.
This does not prevent the user from creating another patch type for
the same image but at least it should not happen by accident.