Fix dragging of water flags in patch editor
If you flag a tile as water in the patch editor, it's supposed to
clear the land flag.  PatchEditorCallbackButtonDown correctly did
this for the tile on which you pressed the mouse button down, but
PatchEditorCallbackMouse failed to do this for the tiles over which
you then dragged.  The bug caused those tiles to be flagged as both
land and water, so that neither land nor water units could enter them.

Fix by moving the flag updates to a new function ChangeFlagsAtMouse
and calling that from both PatchEditorCallbackButtonDown and
PatchEditorCallbackMouse.  This function reads the desired changes
from two bitmask variables that PatchEditorCallbackButtonDown
initializes.  These variables replace the DraggingFlag and
SetFlagWhileDragging flags.

When clearing MapFieldWaterAllowed and setting MapFieldLandAllowed,
also clear MapFieldCoastAllowed and MapFieldNoBuilding.

Fixes: http://savannah.nongnu.org/patch/?7192