Support 1024x600 and 1366x768 resolutions.
List those resolutions in the graphics options screen, and select
the appropriate graphics/ui/ui_*_bpanel.png files for them.
Because all the bpanel graphics have the same width and only the
height varies, choose the right file based on the screen height,
not on the screen width.  This causes some changes:

resolution      old                     new
640x480         ui_640_bpanel.png     = ui_640_bpanel.png
800x600         ui_800_bpanel.png     = ui_800_bpanel.png
1024x600        N/A                     ui_800_bpanel.png
1280x720        ui_1024_bpanel.png    = ui_1024_bpanel.png
1024x768        ui_1024_bpanel.png    = ui_1024_bpanel.png
1366x768        N/A                     ui_1024_bpanel.png
1440x900        ui_1600_bpanel.png    > ui_1280_bpanel.png
1280x1024       ui_1024_bpanel.png    < ui_1600_bpanel.png
1680x1050       ui_1600_bpanel.png    = ui_1600_bpanel.png
1920x1080       ui_1600_bpanel.png    = ui_1600_bpanel.png
1600x1200       ui_1600_bpanel.png    = ui_1600_bpanel.png
1920x1200       ui_1600_bpanel.png    = ui_1600_bpanel.png

The panels on 1440x900 and 1280x1024 now look better than before.

If the resolution does not have an exactly matching bpanel graphic,
then choose the next bigger one, so that it covers the whole area
and the game doesn't crash trying to load a missing graphic file.