Add GCC __attribute__ to printf-like functions
Define a new macro GCC_ATTR, which expands to __attribute__ if using GCC
and to nothing otherwise.

Add to each printf-like function an attribute
GCC_ATTR((format(printf, FORMAT-STRING-ARG, FIRST-FORMATTED-ARG)))
so that GCC will check that the types of the formatted arguments
match the format specifiers in the format string.
(These checks revealed the bug previously fixed in SaveUnit.)